Perltidy improvements
The `bin/run-perltidy.pl` script is set up the same as the corresponding
webwork2 script to autodetect the pg root directory and no longer
depends on the PG_ROOT environment variable.

A new perltidyrc file (`bin/perltidy-pg.rc`) is added.  This is used for
perl tidying pg code.  The only differences are that spaces are used for
indentation, and a line limit of 80 characters is used instead of 120.

The `WeBWorK::PG::Tidy` module is added.  This module is a wrapper
around the `Perl::Tidy` module that basically does everything the
`bin/perltidy-pg.pl` script did before.  That script now just calls the
exported `pgtidy` method.  This module will also be used by the webwork2
PG problem editor to tidy PG problem code.