Fix line numbers in warnings from problems and macros.
If you have edited problem files and received warnings, you are probably
familiar with the fact that the line numbers are off.  The reason for
this is new lines added to the evaluated strings.  This just removes
those new lines.  See the comment on line 23 of WWSafe.pm for the
motivation for this fix.