Clean up the PG.pl POD.
All POD in the file is moved up to where each method is defined.  This
ensures that the order of the generated html is consistent with the
order in the file, and makes it easier to maintain the POD.  Methods are
now `head2`'s instead of items so that they are in the index.