Clean up the pretty_print method in PGUtil.pm
Currently there are some issues with the method. Undefined variables
used in strings cause warnings, and a string containing the word "hash"
can cause errors. Furthermore, the formatting of the output is quite bad
and uses invalid html (for its HTML display mode).

Also make it possible to override all of the arguments for the
`PGUtil.pm` `pretty_print` method when calling the `PG.pl` pretty_print
method. Thus you can call `pretty_print(output, 'text', 10)` instead of
`$PG->pretty_print(output, 'text', 10)` for this anymore. This is useful
in debugging when you want to see deeper into a MathObject object. The
method still works the same when the additional arguments aren't passed.

I debated on rearranging the last two arguments, but left them for now.
The level is the argument that you would typically want to change. It is
less likely that you would want to change the display mode.