Add a text post processing phase to the translator.
Macros or problems can add a hook by calling
`add_content_post_processor` with a subroutine that will be called after
the translator has processed answers.  For all display modes except TeX,
the subroutine will be passed the problem text parsed into a Mojo::DOM
object and a reference to the header text.  For the TeX display mode a
reference to the problem text is passed.  The subroutine can then modify
the problem DOM or text, and header text.  The resulting Mojo::DOM
object will be converted back to a string and the problem text reference
returned by the translator set to point to that string.

Note that the `$PG_PROBLEM_TEXT_ARRAY_REF` has been removed from the
translator code.  It is never used and is redundant with the
`$PG_PROBLEM_TEXT_REF` which is all that is used.  It was becoming
annoying to maintain both in a compatible manner.

The hidden MathQuill inputs for the latex string is added in this post
processing stage (see the ENDDOCUMENT method in PG.pl).