Drop POST_HEADER_TEXT support.
There is no valid reason for a problem to use `POST_HEADER_TEXT`.
Anything that you want after the HTML `head` can be added to the problem
text itself.  There was an invalid comment about iframes suggesting that
this method be used if a page embedded in an iframe contains a form.
However, it does not need to be outside of the problem form.  It is in
an iframe which is a completely separate page, and will not conflict
with the outer page's form.

Note that a problem author can still add to the `head` section of the
HTML page with `HEADER_TEXT`.

This method is not used by any problem in the OPL or Contrib.