Remove html formatting in a warn statement.
Here is a minimal working example to demonstrate the change:
```perl
DOCUMENT();
loadMacros('PGstandard.pl');
BEGIN_TEXT
\{ foo() \}
END_TEXT
ENDDOCUMENT();
```
Also remove the `$PAR` around the output when these errors occur. This
makes the output look more like the code to begin with.

With the develop branch the warning messages for this problem will be:

* `ERROR in old_safe_ev, PGbasicmacros.pl: <PRE>`
* `## There is an error occuring inside evaluation brackets \{ ...code... \}`
* `## somewhere in an EV2 or EV3 or BEGIN_TEXT block.`
* `## Code evaluated:`
* `## foo()`
* `##Undefined subroutine &main::foo called at line 1 of (eval 7649).`
* `##</PRE><BR>`
* `at line 2080 of [PG]/macros/core/PGbasicmacros.pl.`

Since these warnings are now xml escaped by webwork2 the html formatting
no longer works.

With this branch they will be:

* `There is an error occuring inside evaluation brackets \{ ...code... \}`
* `somewhere in an EV2, EV3, or BEGIN_TEXT block.`
* `Code evaluated:`
* `foo()`
* `Errors:`
* `Undefined subroutine &main::foo called at line 1 of (eval 7742).`

I don't think that the "ERROR in old_safe_ev, PGbasicmacros.pl" part is
useful for problem authors, so that was dropped.

Also remove the `;` from the end of the eval, and the comment about it.
Testing shows that comment is false.