Update the parserWordCompletion.pl macro.
Use html native data list element for word completion instead of a
jquery-ui autocomplete.

The message that was previously shown in a native html dialog if a
student entered an answer that is not among the given choices is instead
shown as a feedback message in the attempts table.

With these changes no javascript or css is even needed for the macro
anymore.  However, there may be accessibility concerns.  See the
[accessibility concernts](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/datalist#accessibility_concerns)
on the MDN web docs on this.  Although, I doubt that the jquery-ui
autocomplete was much better.  In any case, we could try to find another
javascript library for this if we want to try to do better.