Remove the `usePolling` chokidar option in generate-assets.js.
With node version 20 the `usePolling` option is resulting in high cpu
usage. PG does not have any symlinks that need to be followed, so this
option isn't even needed.

Interestingly enough, the PG generate-assets.js script is the one that
actually has high cpu usage with this option, not the webwork2
generate-assets.js script.  Something to do with links not actually
existing?