Add feedback in the post processing phase which replaces the webwork2 attempts table.
A feedback button is placed somewhere in the DOM for each response
group.  Default locations that should work well in most cases are
determined in post processing.  However, there is a
ww-feedback-container css class that is provided in problem.scss, and a
problem author can use it to override the default location by wrapping
the responses in a div or span with that class.

The classes "align-middle" and "align-bottom" can be added to the
container in addition to align the feedback button.  By default the
button is aligned at the top.

The result summary is also generated by PG and is returned in the result
summary. The caller can decide to use it or not, and where to place it.