Fix an issue that makes unanswered CheckboxList answers in a MultiAnswer answered.
When a CheckboxList is one of the answers for a MultiAnswer object and
`allowBlankAnswers` is 1 for the MultiAnswer object, then for some not
quite determined reason the cmp_preprocess method of
parserCheckboxList.pl is called at an earlier stage than otherwise, and
additionally at that point the answer hash `student_value` is a
Value::String whose data array contains one empty element.  As such the
the answer strings end up begin set to the last checkbox answer label
due to the way that the `labelText` method works.

Note that the checker still works correctly, but what is displayed in
the results table is misleading.

Essentially the CheckboxList cmp_preprocess is not correctly detecting
when no checkboxes are checked.  This skips setting answer strings in
the case that no checkboxes are checked.