Remove `be_strict`.
All modules should call `use strict` and certainly have no need of this
method.  They are not restricted from using `use` as they are outside of
the safe zone.  They should also all call `use warnings`.  More work is
needed because many of the modules actually have problems that cause
warnings.

Macros can (and should) call `BEGIN { strict->import }`. This is
actually entirely equivalent to previously calling `BEGIN { be_strict() }`
(due to an incorrect `import` call -- see below) and can be done in the
safe compartment. Note that `use strict` is (essentially) the equivalent
of `BEGIN { require strict; strict->import; }`.  For macros there is no
need to `require strict` as `strict` will already have been loaded. So
using `BEGIN { strict->import }` does essentially what `use strict` does
for a macro.

This does not go as far as is really needed.  All macros should call
`BEGIN { strict->import; }`.  This should be a requirement for macros in
PG going forward.  This is added in `parserGraphTool.pl` in this pull
request, and the one issue with that macro is fixed.  This is also added
in `parserRadioMultiAnswer.pl` which has no issues with it. This is
deferred for other macros.

Also remove the unused `ww_strict.pm` module.  Note that module was not
actually used even by the `be_strict` method.  The `be_strict` method
was defined by

```perl
sub be_strict {
	require ww_strict;
	strict::import();
	return;
}

```

This means that the `ww_strict` module was required, but the line
`strict::import()` did not import that module.  Instead that imports the
real perl `strict` module. In order to import the `ww_strict` module it
would have needed to call `ww_strict->import`. Notice that the
`be_strict` method also used an incorrect way of calling the `import`
method of `strict`. The `import` method is a package method that should
be called with an invocant (via `strict->import`, and not a static
method which would be called as `strict::import()` as was done in the
`be_strict` method.  It worked though because the `import` method
doesn't use the invocant, and the `be_strict` method did not accept any
arguments.