Improvements to the image view dialog.
Fix aria labels.  Those can't be set with a camel case object attribute.

Make focus behavior work correctly for the zoom in and out buttons.

The arrow keys now move the dialog left, right, up, or down 10 pixels.
If the control key is down with an arrow key, then it moves 50 pixels.
If the shift key is down instead, then it moves only 1 pixel.

Note that the imageview.css file is converted to scss in the
imageview.scss file.

The imageview.js file has been prettified.