Alternate to #946 and probably better (slightly) approach.
Investigating things closer what I see is that the `Real` call added in
the problem in question, that context happens to be a variant of the
`Fraction-NoDecimals` context.  The `Real` call that was added in #935
is not creating a fraction in that context when it should.

This alternate approach only returns the result of the `Real` call in
the case that an ambiguous value is detected.  Otherwise it continues on
the the rest of the compute method the way that it did before #935 was
added.