Restructure how answer names are associated with answer evaluators.
Before describing the restructuring some terminology needs to be
clarified.

Henceforth we will no longer refer to any answer as being "unlabeled".
All answers are labeled (or better "named" -- more on that later), and
always have been.  Every answer rule method calls a named answer rule
method, and `ANS` calls `NAMED_ANS` (previously it called
`LABELED_ANS`).  Instead answers are either explicitly named (meaning an
author chose the name -- hopefully by calling `NEW_ANS_NAME`), or
implicitly named (meaning that the association of the name and evaluator
is handled by PG by order of ANS and ans_rule calls).

Also, the term "label" should not be used for this. In HTML this is the
"name" attribute on the input, and a "label" is a different thing
entirely. As such the deprecation flag on the `NAMED_ANS` method has
been removed.  In fact, `NAMED_ANS` is no longer an alias for
`LABELED_ANS`.  Instead it is the other way around.  So we should
encourage using `named_ans_rule` methods together with `NAMED_ANS`.

Now for the restructuring.  Instead of using a fragile pairing of answer
name and answer evaluator counts, two FIFO stacks and a name answer hash
are used.  The stacks are basically as described by @dpvc in
https://github.com/openwebwork/pg/issues/944#issuecomment-1781881663.

When an `ans_rule` method is called is calls `NEW_ANS_NAME` as before.
However, `NEW_ANS_NAME` only increments a counter that is only used to
make the answer names unique, and does not record the answer anymore.
The `ans_rule` method is now responsible for recording the answer after
calling `NEW_ANS_NAME`.  It does this by calling
`RECORD_IMPLICIT_ANS_NAME`.  That method pushes the answer name onto the
implicit answer name stack.  Then the named answer rule method (which
must be called by the answer rule method) records the answer by calling
`RECORD_ANS_NAME`.  That actually adds the answer name to the
`PG_ANSWERS_HASH`.  Then when `ANS` is called it shifts the first answer
name from the stack and associates that answer name to the given answer
evaluator by calling `NAMED_ANS`.

The previous paragraph is assuming that the answer rule method is called
before the `ANS` method.  If the `ANS` method is called first, then it
pushes the answer evaluator onto the other stack which is the implicit
evaluator stack.  Then when the `ans_rule` method is called it shifts
the first evaluator from the stack and associates it with the given
answer name by calling `NAMED_ANS` when the answer is recorded by the
named answer rule method.

In the case that an author provides a name (which should be obtained by
calling `NEW_ANS_NAME` to obtain a name that will work in all scenarios
in which problems are rendered), then these explicit names and
associated evaluators are not added to the above stacks.  Of course this
is done as before by calling a name answer rule method and `NAMED_ANS`
(or its alias `LABELED_ANS`).  This basically circumvents the stacks and
makes the association directly.  If the `NAMED_ANS` method is called
first, then the given name and evaluator are added to the explicit named
evaluator hash.  Then when the named answer rule method is called with
the same answer name it removes the evaluator from the hash and adds it
to the PG_ANSWERS_HASH.  This ensures that the order of PG_ANSWERS_HASH
is the same as the order that the answer rule methods are called.  Thus
the order in the (soon to be removed) attempts table will always be
correct.  This order will still matter somewhat without the attempts
table in some cases as it is the order that the answer evaluators are
called in answer processing.

With this pull request I think that all of the examples described in
work).  With this it is very easy to intersperse implicitly and
explicitly names answers.  If a `MultiAnswer` object is used, its
answers can be mixed with other answers easier.  If PGML is used
(tweaked by this pull request as discussed in #944), then you don't even
need to use the `namedRules` option to do so and PGML handles the
details of getting the order correct.  If PGML is not used, you still
need to use `namedRules`, but you don't need to name the other answers
that are mixed in.

Of course this is completely compatible with existing problems and there
is nothing changed for problem authors in terms of the PG answer rule
API.

Now it is truly valid to tell authors to use `NEW_ANS_NAME` to get an
answer name to use, and it won't interfere with the association of
answer names to answer evaluators.  However, care should be taken to
call the `NEW_ANS_NAME` method with the answer rule method it will be
used for.  This is because the answer names are used to generate the
aria labels for answers.  If the `NEW_ANS_NAME` is called before or
after another answer rule method then the answer names will be out of
order.  You will have something like AnSwEr0002, AnSwEr0001, AnSwEr0003
for order of the names used, and thus the aria labels will be 'answer
2', 'answer 1', 'answer 3' in the order of the page.

There was also POD clean up in PGcore.pm, and some reordering of the
methods therein to make the POD headers work better.