Add display "inline" to the feedback popover.
When a feedback button is clicked the feedback popover is injected into
the DOM in its parent element.  A bootstrap popover is constructed with
`div`s.  As such if the input the feedback button is associated with has
content after it in the same line, the injected display block `div`
causes the line to break.

This adds `display:inline` to the popover to prevent the line break.

Unfortunately, this does result in invalid html, but it works.  The only
fix that I can think of is to change the `span`s added around the inputs
into `div`s that also have `display:inline` set.  To fix problems that
do not use PGML and use `$PAR` for line breaks, the `$PAR` definition
would need to be changed from being a `p` tag to being a `div` with a
margin (same as PGML does).

Also set a minimum width on the popover of 200px.  Now that the popover
is contained in the parent of the feedback button if that parent
is narrow, then the popover width is also quite narrow.  This looks
better in that case.