Update the LiveGraphics3D.pl macro with modern JavaScript.
This removes the dependence on jQuery.

Unfortunately, the npm installed version of x3dom does not include the
x3dom-full.js file, and the live graphics macro uses x3dom elements
(like Polyline2D) which are not supported by the basic x3dom.js file. So
that library can not be installed via npm. So instead the x3dom files
are served from the www.x3dom.org site.  Another option would be to go
back to the old school approach of including the files in the repository
directly.  Note that the npm package that we were using is made
available by a third party. The main thing that you will notice from
switching to using the x3dom-full.js file is that now the cube
containing the 3D graph will be visible again (that uses the
Polyline2D).

Also the version of x3dom is upgraded to 1.8.3.

Also, the underlying html injected by Perl in macro is different.
Instead of using a table with a single row and cell, a div is used.
That table actually worked quite poorly in many cases, and certainly was
not the right thing to use.  Furthermore, there is no injected script
element.  Instead the options are passed via data attributes on that
div.