Change from diplay:inline to position:absolute for the feedback popover.
I suspect that bootstrap adds position:absolute to the element after the
popover has already been inserted into the DOM, and that is what is
causing the line breaking in Safari.  So this adds position:absolute to
the `ww-feedback-popover` class so that the popover already is
absolutely positioned.  I don't think that display:inline is needed
since technically an absolutely positioned div can not be display
inline..