Warn if a macro file has a POD directive on the first line but don't fail.
This adds a new line to the beginning of the code for a macro file that
starts with `=` which will usually mean an invalid POD directive on the
first line of the file (and is invalid in any case for the first
character of a Perl file).

In that case a warning is also shown.  This warning will only be shown
if the view_problem_debugging_info flag is set (this corresponds to the
webwork2 permission that ta's and above have).

This is to address issue #961.