Update the methods for parsing and interpreting the LiveGraphics3D data.
This parses and interprets the data more like I believe it is intended
to be parsed according to what I have read in the limited documentation
that is available.

This also honors a several of the commands that were not previously
honored, but are used in the various LiveGraphics3D macros.  As a result
some problems in the OPL will look a bit different.  In most cases the
options do improve appearance.