Add a prefix to the `align-middle` and `align-bottom` feedback container css classes.
Using `align-middle` and `align-bottom` can conflict with the bootstrap
classes with the same name.  The feedback alignment classes are
different in that they align the feedback button relative to the
feedback container, and the bootstrap classes align the container
itself.

Note that these classes are not used internally by pg (at this point),
but are provided for authors using MultiAnswer's with `singleResult`
true that want to position the feedback button relative to the feedback
container.  Also note that there is no `align-top` because that is the
default behavior for the `ww-feedback-btn`.