Don't always show "Ungraded" for essay problems.
This requires a change to webwork2 (or any other front end).  Basically,
the front end needs to let PG know if the problem has been graded or not
via the environment "needs_grading" flag.

The feedback processing in the ENDDOCUMENT method of PG.pl and
PGessaymacros.pl use this flag to show the correct messages for essay
answers.

Note that the methods in PG.pl now has more generic code for handling
answers like those from the PGessaymacros.pl macro instead of being
specifically for that macro.  This done with the `manuallyGraded` and
`needsGrading` options that can be set by a `feedback_options` method
for a custom answer type.

I think that the answer messages and feedback popover titles need
further examination.  I am not sold on using "Graded" and "Ungraded" for
manually graded problems.  I also don't like the messages that state
things like "All of the gradable answers are correct".  In reality all
questions are "gradable", it is just a matter of who or what does the
grading (and perhaps when that grading is done). I am open to
suggestions for improvements to these messages.