For PGML escape all characters (\, `, and $) in a line of html output not just the first.
All of @somiaj's suggested escape code comments in #1980 contained the
`g` flag on the regular expression, but when #878 was actually
committed, the code did not.  As a result only the first character on
the line will actually be escaped.  This results in rather unexpected
MathJax rendering.  For example, if the line contains
```
`x + y = 3` and `x - y = 4`
```
then this is rendered as
```
<span class="text2jax_ignore">`</span>x + y = 3` and `x - y = 4`
```
Which results in `` ` and ` `` being rendered by MathJax.