Fix warnings for improper PGML block termination.
This is basically @dpvc's suggestion at
https://github.com/openwebwork/pg/issues/654#issuecomment-1843553494.

There were two lines of code changed in his suggestion that seem to be
unrelated.

The first adds a second return value which is the "inline" math mode to
the list returned on line 1358.  I also made that change, since it seems
correct.  The other return from the same method also returns a list with
two values, the second of which is a math mode.

The second just changed the parentheses on line 1364.  I just removed
the parentheses because neither the original parentheses or the way they
were changed in the suggestion are needed.