Add an option to show the correct answer with a reveal button.
Technically this does not add an option, but changes how the existing
showCorrectAnswers option in the PG environment works.  Previously it
was purely boolean.  Now it is numeric.

If it is zero, then correct answers are not shown (as before).

If it is 1, then correct answers are shown but hidden, and a "Reveal"
button is shown at first. If that button is clicked, then the answer is
shown.

If set to 2, then correct answers are shown immediately.