Minor clean up and issue fixing for scaffold.pl.
The code for hiding (or really dimming) the results for unnopened
sections in the attempts table is no longer relevant.  So that is
removed.

Also add `BEGIN { strict->import }` and fix the issues it reveals.
This issue on line 317 is my fault.  In a previous pull request I moved
that line up before `$self` was defined to group the error calls
together at the beginning, but missed that `$self` needed to be changed
to `$scaffold` to do that (which is what `$self` is later set to be).