Add prettier formatting of JavaScript, style, and HTML files, and check formatting in a workflow.
Prettier is used to format JavaScript, style, and HTML files in htdocs.
The configuration chosen is in .prettierrc and .editorconfig.  Prettier
doesn't really have many options as it is very opinionated.  I am not
entirely happy with everything prettier does, but it does give a more
uniform code format to files.  Furthermore, it is nice to be able to be
able to just run prettier and not need to worry about formatting code.

A workflow now checks formatting of JavaScript, style, and HTML files
with prettier in addition checking formatting of perl files with
perltidy.  Note the "linter" workflow has been renamed to
"check-formats", because that is what it does, and this really does not
have anything to do with linting.

Note that developers can (and should) run `npm run prettier-format` in
the htdocs directory to format files.  You can also run `npm run
prettier-check` to check that formatting of these files is correct, and
not actually change the files.  This is what the workflow does.

The file conf/pg_config.dist.yml is also formatted and checked with
prettier.